feat(resume-onload): simpifies interface by passing settings through, but also removes JobFunc support
